[core] Attribute: Add notion of depth for attributes with parents
For attributes in `GroupAttribute` and `ListAttribute`, the notion of
parent attribute exists through the `root` property. As a parent can
itself have a parent, the `depth` property describes how many levels
there are between the attribute and the root level.

A value of 0 means that the attribute is at the root level, and it
increases as it gets deeper.